+++ www/docs/server/subscriptions.md
@@ -147,8 +147,8 @@ Since subscriptions are async iterators, you have to go through the iterator to 
 
 ### Example with zod
 
-```ts title="zAsyncIterable.ts"
-import { isTrackedEnvelope, tracked, type TrackedEnvelope } from '@trpc/server';
+```ts title="zAsyncGenerator.ts"import type { TrackedEnvelope } from '@trpc/server';
+import { isTrackedEnvelope, tracked } from '@trpc/server';
 import { z } from 'zod';
 
 function isAsyncIterable<TValue, TReturn = unknown>(
@@ -156,46 +156,72 @@ function isAsyncIterable<TValue, TReturn = unknown>(
 ): value is AsyncIterable<TValue, TReturn> {
   return !!value && typeof value === 'object' && Symbol.asyncIterator in value;
 }
+const trackedEnvelopeSchema =
+  z.custom<TrackedEnvelope<unknown>>(isTrackedEnvelope);
 
 /**
- * Zod schema for an async iterable
+ * Zod schema for an async generators
  * - validates that the value is an async iterable
- * - parses each item in the async iterable
+ * - validates each item in the async iterable
+ * - validates the return value of the async iterable
  */
-export function zAsyncIterable<TYieldIn, TYieldOut>(
-  yieldSchema: z.ZodType<TYieldIn, any, TYieldOut>,
-) {
+export function zAsyncGenerator<
+  TYieldIn,
+  TYieldOut,
+  TReturnIn = void,
+  TReturnOut = void,
+  Tracked extends boolean = false,
+>(opts: {
+  /**
+   * Validate the value yielded by the async generator
+   */
+  yield: z.ZodType<TYieldIn, any, TYieldOut>;
+  /**
+   * Validate the return value of the async generator
+   * @remark not applicable for subscriptions
+   */
+  return?: z.ZodType<TReturnIn, any, TReturnOut>;
+  /**
+   * Whether if the yielded values are tracked
+   * @remark only applicable for subscriptions
+   */
+  tracked?: Tracked;
+}) {
   return z
-    .custom<AsyncIterable<TYieldIn, void, unknown>>((val) =>
-      isAsyncIterable(val),
-    )
+    .custom<
+      AsyncGenerator<
+        Tracked extends true ? TrackedEnvelope<TYieldIn> : TYieldIn,
+        TReturnIn
+      >
+    >((val) => isAsyncIterable(val))
     .transform(async function* (iter) {
-      for await (const data of iter) {
-        yield yieldSchema.parseAsync(data);
+      const iterator = iter[Symbol.asyncIterator]();
+      let next;
+      while ((next = await iterator.next()) && !next.done) {
+        if (opts.tracked) {
+          const [id, data] = trackedEnvelopeSchema.parse(next.value);
+          yield tracked(id, await opts.yield.parseAsync(data));
+          continue;
+        }
+        yield opts.yield.parseAsync(next.value);
       }
-    });
-}
-
-/**
- * Zod schema for an async iterable
- * - Validates that the schema is tracked
- * - Validates that the value is an async iterable
- * - Validates each item in the async iterable
- */
-export function zAsyncIterableTracked<TYieldIn, TYieldOut>(
-  yieldSchema: z.ZodType<TYieldIn, any, TYieldOut>,
-) {
-  const trackedEnvelopeSchema =
-    z.custom<TrackedEnvelope<TYieldIn>>(isTrackedEnvelope);
-
-  return z
-    .custom<AsyncIterable<TrackedEnvelope<TYieldIn>, any, any>>(isAsyncIterable)
-    .transform(async function* (iter) {
-      for await (const data of iter) {
-        const [id, value] = trackedEnvelopeSchema.parse(data);
-        yield tracked(id, yieldSchema.parse(value));
+      if (opts.return) {
+        return await opts.return.parseAsync(next.value);
       }
-    });
+      return;
+    }) as any as z.ZodType<
+    AsyncGenerator<
+      Tracked extends true ? TrackedEnvelope<TYieldIn> : TYieldIn,
+      TReturnIn,
+      unknown
+    >,
+    any,
+    AsyncGenerator<
+      Tracked extends true ? TrackedEnvelope<TYieldOut> : TYieldOut,
+      TReturnOut,
+      unknown
+    >
+  >;
 }
 ```
 
@@ -203,14 +229,23 @@ Now you can use this helper to validate the output of your subscription procedur
 
 ```ts title="_app.ts"
 import { publicProcedure, router } from '../trpc';
-import { zAsyncIterable } from './zAsyncIterable';
+import { zAsyncGenerator } from './zAsyncGenerator';
 
 export const appRouter = router({
   mySubscription: publicProcedure
-    .output(zAsyncIterable(z.number()))
+    .output(
+      zAsyncGenerator({
+        yield: z.number(),
+        tracked: true,
+      }),
+    )
     .subscription(async function* (opts) {
+      let index = 0;
+      if (opts.input.lastEventId) {
+        index = Number(opts.input.lastEventId);
+      }
       while (true) {
-        yield Math.random();
+        yield ++i;
 
         await new Promise((resolve) => setTimeout(resolve, 1000));
       }
